added read_as_array showing progress bar
