+++ earth_data_kit/xarray_boosted/__init__.py
@@ -1,2 +1,2 @@
-import earth_data_kit.xarray_boosted.plot
+import earth_data_kit.xarray_boosted.edk_accessor
 import earth_data_kit.xarray_boosted.entrypoint
